[#4] refactor: bcrypt를 사용한 hash 관련 모듈을 helper에서 util로 이동
