+++ src/services/user.ts
@@ -5,7 +5,7 @@ import UserEntity from '../entity/user';
 import UserRepository from '../repositories/user';
 import ErrorResponse from '../utils/error-response';
 import { commonError } from '../constants/error';
-import * as hashHelper from '../helpers/hash';
+import { generateHash } from '../utils/hash';
 
 @Service()
 class UserService {
@@ -40,7 +40,7 @@ class UserService {
         throw new ErrorResponse(commonError.conflict);
       }
 
-      const hashedPassword = hashHelper.generateHash(password);
+      const hashedPassword = generateHash(password);
       const createdUser = await this.userRepository.addItem(id, hashedPassword);
       const { idx, createdAt, updatedAt } = createdUser;
       return { idx, createdAt, updatedAt };
