+++ src/utils/hash.ts
